+++ module/CSV_Procurements_extractor/src/main/java/cz/cuni/mff/xrg/odcs/procurementExtractor/datanest/AbstractDatanestHarvester.java
@@ -1,11 +1,4 @@
-package cz.cuni.mff.xrg.odcs.extractor.datanest;
-
-
-import au.com.bytecode.opencsv.CSVReader;
-import cz.cuni.mff.xrg.odcs.extractor.data.AbstractRecord;
-import cz.cuni.mff.xrg.odcs.extractor.file.CsvProcurementsExtractor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+package cz.cuni.mff.xrg.odcs.procurementExtractor.datanest;
 
 import java.io.BufferedReader;
 import java.io.File;
@@ -19,20 +12,26 @@ import java.util.Calendar;
 import java.util.Properties;
 import java.util.Vector;
 
-public abstract class AbstractDatanestHarvester<RecordType extends AbstractRecord> extends AbstractHarvester<RecordType> {
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
+import au.com.bytecode.opencsv.CSVReader;
+import cz.cuni.mff.xrg.odcs.procurementExtractor.data.AbstractRecord;
+import cz.cuni.mff.xrg.odcs.procurementExtractor.core.CsvProcurementsExtractor;
+
+public abstract class AbstractDatanestHarvester<RecordType extends AbstractRecord> extends AbstractHarvester<RecordType> {
 
     public final static String DATANEST_DATE_FORMAT = "yyyy-MM-dd";
     protected final static SimpleDateFormat sdf = new SimpleDateFormat(DATANEST_DATE_FORMAT);
     private static Logger LOG = LoggerFactory.getLogger(AbstractDatanestHarvester.class);
 
-
     /**
-     * @param sourceUrlKey key used to get the source URL from Datanest properties
-     * @throws java.io.IOException when initialization of primary repository fails
+     * @param sourceUrlKey
+     *            key used to get the source URL from Datanest properties
+     * @throws java.io.IOException
+     *             when initialization of primary repository fails
      */
-    public AbstractDatanestHarvester(String sourceUrlKey)
-          {
+    public AbstractDatanestHarvester(String sourceUrlKey) {
 
         super();
 
@@ -41,11 +40,10 @@ public abstract class AbstractDatanestHarvester<RecordType extends AbstractRecor
     abstract public RecordType scrapOneRecord(String[] row) throws ParseException;
 
     /**
-     * Most common implementation of harvesting code in our current Datanest
-     * harvesters.
-     *
-     * @param sourceFile temporary file holding freshly obtained data to harvest from
-     *                   when some repository error occurs
+     * Most common implementation of harvesting code in our current Datanest harvesters.
+     * 
+     * @param sourceFile
+     *            temporary file holding freshly obtained data to harvest from when some repository error occurs
      */
     @Override
     public Vector<RecordType> performEtl(File sourceFile) throws Exception {
@@ -53,18 +51,16 @@ public abstract class AbstractDatanestHarvester<RecordType extends AbstractRecor
         Vector<RecordType> records = new Vector<RecordType>();
         try {
             // "open" the CSV dump
-            CSVReader csvReader = new CSVReader(new BufferedReader(
-                    new FileReader(sourceFile)));
+            CSVReader csvReader = new CSVReader(new BufferedReader(new FileReader(sourceFile)));
             // TODO: If we store also the original copy of the data (say in
             // Jacrabbit) and perform a "diff" on that and previous version we can:
             // a) determine also removed records (which current implementation
-            //    does not know to do)
+            // does not know to do)
             // b) determine new and updated records without downloading records
-            //    for all IDs ...
+            // for all IDs ...
             // c) ... instead noting only changed records in say vectors and
-            //    processing only those (thus saving a LOT of resources assuming
-            //    changes and additions are small and infrequent)
-
+            // processing only those (thus saving a LOT of resources assuming
+            // changes and additions are small and infrequent)
 
             // TODO: check the header - for now we simply skip it
             csvReader.readNext();
@@ -76,9 +72,9 @@ public abstract class AbstractDatanestHarvester<RecordType extends AbstractRecor
             Properties prop = new Properties();
 
             try {
-                //load a properties file from class path, inside static method
+                // load a properties file from class path, inside static method
                 prop.load(CsvProcurementsExtractor.class.getClassLoader().getResourceAsStream("config.properties"));
-                //get the property value and print it out
+                // get the property value and print it out
                 debugProcessOnlyNItems = Integer.valueOf(prop.getProperty("debugProcessOnlyNItems"));
 
             } catch (IOException e) {
@@ -86,7 +82,6 @@ public abstract class AbstractDatanestHarvester<RecordType extends AbstractRecor
             }
 
             LOG.debug("value of debugProcessOnlyNItems: " + debugProcessOnlyNItems);
-
 
             // read the rows
             String[] row;
@@ -105,16 +100,13 @@ public abstract class AbstractDatanestHarvester<RecordType extends AbstractRecor
                     // prematurely - this will cause last fetched line of CSV to
                     // be incomplete
                     LOG.warn("index out of bound exception (broken connection?)", e);
-                    LOG.warn("skipping following record: "
-                            + Arrays.deepToString(row));
+                    LOG.warn("skipping following record: " + Arrays.deepToString(row));
                 } catch (ParseException e) {
                     LOG.warn("parse exception", e);
-                    LOG.warn("skipping following record: "
-                            + Arrays.deepToString(row));
+                    LOG.warn("skipping following record: " + Arrays.deepToString(row));
                 }
 
-                if (debugProcessOnlyNItems > 0 &&
-                        recordCounter >= debugProcessOnlyNItems)
+                if (debugProcessOnlyNItems > 0 && recordCounter >= debugProcessOnlyNItems)
                     break;
             }
 
@@ -132,6 +124,5 @@ public abstract class AbstractDatanestHarvester<RecordType extends AbstractRecor
         }
         return records;
     }
-
 
 }
