+++ module/CSV_Procurements_extractor/src/main/java/cz/cuni/mff/xrg/odcs/procurementExtractor/datanest/AbstractHarvester.java
@@ -1,4 +1,4 @@
-package cz.cuni.mff.xrg.odcs.extractor.datanest;
+package cz.cuni.mff.xrg.odcs.procurementExtractor.datanest;
 
 /* Copyright (C) 2011 Peter Hanecak <hanecak@opendata.sk>
  *
@@ -18,40 +18,31 @@ package cz.cuni.mff.xrg.odcs.extractor.datanest;
  * along with Open Data Node.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import cz.cuni.mff.xrg.odcs.extractor.data.AbstractRecord;
-import cz.cuni.mff.xrg.odcs.extractor.serialization.AbstractSerializer;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import java.io.File;
 import java.net.URL;
 import java.util.List;
 import java.util.Vector;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import cz.cuni.mff.xrg.odcs.procurementExtractor.data.AbstractRecord;
+import cz.cuni.mff.xrg.odcs.procurementExtractor.serialization.AbstractSerializer;
+
 /**
- * Stuff common to all harvesters. Harvesters (along with serializers) perform
- * ETL (http://en.wikipedia.org/wiki/Extract,_transform,_load).
+ * Stuff common to all harvesters. Harvesters (along with serializers) perform ETL (http://en.wikipedia.org/wiki/Extract,_transform,_load).
  * <p/>
- * What is a common harvesting work-flow:
- * 1) download the original document(s) from the main source of data
- * 2) store the copy of original(s) in primary repository
- * why: to have audit trail, our own copy, possibly to run next steps
- * from this own copy instead of downloading (possibly unchanged)
- * document repeatedly
- * 3) extract data
- * 4) enhance data: clean-up, correction, correlation, possibly production
- * of new data via math/logic, ...
- * 5) serialize data into format(s) suitable for ODN back-end repository(ies)
- * 6) store the data into ODN back-end repository(ies)
+ * What is a common harvesting work-flow: 1) download the original document(s) from the main source of data 2) store the copy of original(s) in primary
+ * repository why: to have audit trail, our own copy, possibly to run next steps from this own copy instead of downloading (possibly unchanged) document
+ * repeatedly 3) extract data 4) enhance data: clean-up, correction, correlation, possibly production of new data via math/logic, ... 5) serialize data into
+ * format(s) suitable for ODN back-end repository(ies) 6) store the data into ODN back-end repository(ies)
  * <p/>
- * Note in regards to "primary repository": After we harvest the data, we store
- * it into multiple repositories to serve for multiple purposes. In current
- * architecture, it means Jackrabbit as primary document store (with full data:
- * original record, harvested and enhanced record, ...) and secondary stores
- * SOLR (for full-text search) and Sesame (for RDF and SPARQL).
- *
- * @param <RecordType> type of individual record into which the harvested data are stored
- *                     into
+ * Note in regards to "primary repository": After we harvest the data, we store it into multiple repositories to serve for multiple purposes. In current
+ * architecture, it means Jackrabbit as primary document store (with full data: original record, harvested and enhanced record, ...) and secondary stores SOLR
+ * (for full-text search) and Sesame (for RDF and SPARQL).
+ * 
+ * @param <RecordType>
+ *            type of individual record into which the harvested data are stored into
  */
 public abstract class AbstractHarvester<RecordType extends AbstractRecord> {
 
@@ -63,16 +54,17 @@ public abstract class AbstractHarvester<RecordType extends AbstractRecord> {
     private Object primaryRepository = null;
 
     /**
-     * @throws java.io.IOException when initialization of primary repository fails
+     * @throws java.io.IOException
+     *             when initialization of primary repository fails
      */
     public AbstractHarvester() {
 
-        // TODO remove this          primaryRepository
+        // TODO remove this primaryRepository
         this.primaryRepository = null;
         this.serializers = new Vector<AbstractSerializer<RecordType, ?, ?>>();
     }
 
-    // TODO remove this       primaryRepository
+    // TODO remove this primaryRepository
     public Object getPrimaryRepository() {
         return primaryRepository;
     }
@@ -99,17 +91,19 @@ public abstract class AbstractHarvester<RecordType extends AbstractRecord> {
 
     /**
      * Extract, transform and load the data.
-     *
-     * @param sourceFile temporary file holding freshly obtained data to harvest from
+     * 
+     * @param sourceFile
+     *            temporary file holding freshly obtained data to harvest from
      */
     abstract public Vector<RecordType> performEtl(File sourceFile) throws Exception;
 
     /**
-     * Loop through all serializers and pass given records to them. Serializers
-     * will serialize the records and store them.
-     *
-     * @param records list of records to serialize and store
-     * @throws IllegalArgumentException if repository with given name does not exists
+     * Loop through all serializers and pass given records to them. Serializers will serialize the records and store them.
+     * 
+     * @param records
+     *            list of records to serialize and store
+     * @throws IllegalArgumentException
+     *             if repository with given name does not exists
      */
     protected void store(List<RecordType> records) throws Exception {
 
@@ -118,7 +112,6 @@ public abstract class AbstractHarvester<RecordType extends AbstractRecord> {
         {
             return;
         }
-
 
         for (AbstractSerializer<RecordType, ?, ?> serializer : serializers)
             serializer.store(records);
