+++ module/CSV_Procurements_extractor/src/test/cz/cuni/mff/xrg/odcs/extractor/Test.java
@@ -2,8 +2,8 @@ package cz.cuni.mff.xrg.odcs.extractor;
 
 import cz.cuni.mff.xrg.odcs.commons.configuration.ConfigException;
 import cz.cuni.mff.xrg.odcs.dpu.test.TestEnvironment;
-import cz.cuni.mff.xrg.odcs.extractor.file.CsvProcurementsExtractor;
-import cz.cuni.mff.xrg.odcs.extractor.file.CsvProcurementsExtractorConfig;
+import cz.cuni.mff.xrg.odcs.procurementExtractor.core.CsvProcurementsExtractor;
+import cz.cuni.mff.xrg.odcs.procurementExtractor.core.CsvProcurementsExtractorConfig;
 import cz.cuni.mff.xrg.odcs.rdf.interfaces.RDFDataUnit;
 
 public class Test {
@@ -21,7 +21,7 @@ public class Test {
             String input = null;
             env.run(extractor);
         } catch (Exception e) {
-            e.printStackTrace();  //To change body of catch statement use File | Settings | File Templates.
+            e.printStackTrace(); // To change body of catch statement use File | Settings | File Templates.
         } finally {
             // release resources
             env.release();
